--- ood/utils.py
@@ -1,13 +1,0 @@
-"""
-Some utilities used in the package.
-For example, the code related to generalized chi-square estimation is one of the
-utilities that is being used.
-"""
-
-# TODO: add some code for the generalized chi-square statistics such as:
-# 1. Calculating the mean
-# 2. Calculating the variance
-# 3. Calculating the k-th moment using the moment generating function
-# 4. Calculating the cdf
-# 5. Calculating the inverse_cdf
-# 6. Calculating the log_cdf
